Move into test/core/parser/test_parser.py
