+++ tests/core/parser/test_parser.py
@@ -5,7 +5,13 @@
 #
 # The full license is in the file COPYING.txt, distributed with this software.
 #------------------------------------------------------------------------------
+import os
+import sys
 import ast
+import enaml
+import pytest
+import traceback
+from textwrap import dedent
 
 
 def validate_ast(py_node, enaml_node, dump_ast=False, offset=0):
@@ -35,3 +41,44 @@ def validate_ast(py_node, enaml_node, dump_ast=False, offset=0):
             validate_ast(n1, enaml_node[i], offset=offset+1)
     else:
         assert py_node == enaml_node
+
+
+def test_syntax_error_traceback_correct_path(tmpdir):
+    """ Test that a syntax error retains the path to the file
+    
+    """
+    test_module_path = os.path.join(tmpdir.strpath, 'view.enaml')
+    
+    with open(os.path.join(tmpdir.strpath, 'test_main.enaml'), 'w') as f:
+        f.write(dedent("""
+        from enaml.widgets.api import Window, Container, Label
+        from view import CustomView
+
+        enamldef MyWindow(Window): main:
+            CustomView:
+                pass
+
+        """))
+    
+    with open(test_module_path, 'w') as f:
+        f.write(dedent("""
+        from enaml.widgets.api import Container, Label
+
+        enamldef CustomLabel(Container):
+            Label # : missing intentionally
+                text = "Hello world"
+
+        """))
+    
+    try:
+        sys.path.append(tmpdir.strpath)
+        with enaml.imports():
+            from test_main import MyWindow
+        assert False, "Should raise a syntax error"
+    except Exception as e:
+        tb = traceback.format_exc()
+        print(tb)
+        lines = tb.strip().split("\n")
+        assert 'File "{}", line 5'.format(test_module_path) in lines[-4]
+    finally:
+        sys.path.remove(tmpdir.strpath)
